The first commit with actual project code.
This will build. It's a highly modified version of the 'specan' spectrum
analyzer project for the IM-ME with radio control, modem, packet and other
settings tuned for pager frequencies and data rates.

Currently it is capable of recognizing the synchronization word of the FLEX
pager message, and is coded to be hard tuned to a particular frequency for
wide area broadcast in the United States.

Currently the code facilitates the confirmation that the right radio control
settings have been configured for the CC1110 device.

Much graditude is owed to Travis Goodspeed (sic?) and others in the IM-ME
hacking community for their development of JTAG programmers, firmwares,
operating systems and tutorials as well as Girltech and TI for the hardware
platforms (radio product and SoC).