Able to pull 0xDEA0 (4-FSK, 3200 baud) FLEX messages from the air and interrupt.
This modification exposes the behavior of register bits via easy-to-
call functions. Some of them still need to be parameterized, and there
remains functionality that has not been functionalized.

The main loop has been modified to read the frame information word,
which is inconveniently in LSB order (the SoC reads MSB). This is used
to extract the cycle and frame number - and thus the current time
(minute and second). Eventually the cycle and frame number can be used
to put the CPU into sleep and wake before the next pager message comes
in, saving a large amount of power.

Extrernally, there is code to compress english strings with an 16-bit
word size arithmetic code. It achieves ~2:1 ratio which is not likely
to be beat and still respect both CPU limitations and crypto concerns.
Eventually this code will make it into the pager system, waiting for
full correctness testing on the CC1110 hardware.

Next steps are to sleep during the Sync2 stage of the FLEX protocol,
and then gather the bits for frames. These bits are interwoven, and
the code will need to reshuffle them for assembly. Around this time
an implementation of BCH codes will be required to correct spurious
bit errors.